Split the servicing ledger test into focused files
The servicing ledger test had grown to 826 lines, well over the 400-line
guideline. Split it into five focused files under the same folder, each named
for what it checks:

- ledger-accounts: cost accounts, listing profit, append-only history
- ledger-idempotency: idempotency keys and replay protection
- ledger-cost-editing: editing costs via correcting adjustments
- ledger-validation: bad input is rejected, never reaches the ledger
- ledger-reader: what the operator sees, plus ledger visibility

The fixtures every file shared moved into a new #test-utils/servicing-ledger.ts,
and the two general non-empty flash-message checks moved to #test-utils/
assertions.ts beside parseFlashCookie. No test behaviour changed — the same
41 tests run, just organised so each file is small enough to read and to
mutation-test quickly.
